reflect: move out reflect-related entities from assistant
