+++ module/core/reflect_tools/src/reflect/wrapper.rs
@@ -1,0 +1,59 @@
+//!
+//! Collection of wrappers.
+//!
+
+/// Internal namespace.
+pub( crate ) mod private
+{
+}
+
+pub mod aref;
+pub mod maybe_as;
+
+#[ doc( inline ) ]
+#[ allow( unused_imports ) ]
+pub use protected::*;
+
+/// Protected namespace of the module.
+pub mod protected
+{
+  #[ doc( inline ) ]
+  #[ allow( unused_imports ) ]
+  pub use super::
+  {
+    aref::orphan::*,
+    maybe_as::orphan::*,
+  };
+}
+
+/// Orphan namespace of the module.
+pub mod orphan
+{
+  #[ doc( inline ) ]
+  #[ allow( unused_imports ) ]
+  pub use super::exposed::*;
+}
+
+/// Exposed namespace of the module.
+pub mod exposed
+{
+  #[ doc( inline ) ]
+  #[ allow( unused_imports ) ]
+  pub use super::
+  {
+    aref::exposed::*,
+    maybe_as::exposed::*,
+  };
+}
+
+/// Prelude to use essentials: `use my_module::prelude::*`.
+pub mod prelude
+{
+  #[ doc( inline ) ]
+  #[ allow( unused_imports ) ]
+  pub use super::
+  {
+    aref::prelude::*,
+    maybe_as::prelude::*,
+  };
+}
