+++ module/core/reflect_tools/tests/inc/group1/only_test/reflect_struct.rs
@@ -12,7 +12,7 @@ fn reflect_basic_test()
 
   a_id!( reflect::reflect( &ins ).is_container(), true );
   a_id!( reflect::reflect( &ins ).len(), 3 );
-  a_id!( reflect::reflect( &ins ).type_name(), "tests::inc::reflect_struct_manual_test::Struct1" );
+  a_id!( reflect::reflect( &ins ).type_name(), "tests::inc::group1::struct_manual_test::Struct1" );
   let names = reflect::reflect( &ins ).elements().map( | e | e.key ).collect::< Vec< _ > >();
   a_id!( names, vec![ reflect::Primitive::str( "f1" ), reflect::Primitive::str( "f2" ), reflect::Primitive::str( "f3" ) ] );
   let types = reflect::reflect( &ins ).elements().map( | e | e.val.type_name() ).collect::< Vec< _ > >();
