+++ module/core/reflect_tools/tests/inc/group1/struct_in_struct_manual_test.rs
