+++ module/core/reflect_tools/tests/inc/group1/struct_with_lifetime_manual_test.rs
